requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2.1
django-cors-headers==4.7.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-yasg==1.21.10
earthengine-api==1.5.17
google-api-core==2.24.2
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
kombu==5.5.3
numpy==2.2.6
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
